+++ src/main/java/de/maxhenkel/camera/integration/jei/JEIPlugin.java
@@ -1,0 +1,23 @@
+package de.maxhenkel.camera.integration.jei;
+
+import de.maxhenkel.camera.Main;
+import de.maxhenkel.camera.gui.AlbumScreen;
+import de.maxhenkel.camera.gui.ImageScreen;
+import mezz.jei.api.IModPlugin;
+import mezz.jei.api.registration.IGuiHandlerRegistration;
+import net.minecraft.util.ResourceLocation;
+
+@mezz.jei.api.JeiPlugin
+public class JEIPlugin implements IModPlugin {
+
+    @Override
+    public ResourceLocation getPluginUid() {
+        return new ResourceLocation(Main.MODID, "camera");
+    }
+
+    @Override
+    public void registerGuiHandlers(IGuiHandlerRegistration registration) {
+        registration.addGuiContainerHandler(ImageScreen.class, new ImageContainerHandler<>());
+        registration.addGuiContainerHandler(AlbumScreen.class, new ImageContainerHandler<>());
+    }
+}